+++ docs/gallery/domain/plot_icephys.py
@@ -3,7 +3,7 @@
 .. _icephys_tutorial_new:
 
 Intracellular electrophysiology
-==================================================================
+===============================
 
 The following tutorial describes storage of intracellular electrophysiology data in NWB.
 NWB supports storage of the time series describing the stimulus and response, information
@@ -16,17 +16,21 @@ experiment.
        tutorials as part of the :incf_collection:`NWB Course <neurodata-without-borders-neurophysiology-nwbn>`
        at the INCF Training Space.
 
-
-To describe the organization of intracellular experiments the metadata is organized
-hierarchically in a sequence of tables. All of the tables are so-called DynamicTables
-enabling users to add columns for custom metadata.
-
 .. figure:: ../../figures/plot_icephys_table_hierarchy.png
     :width: 700
     :alt: Intracellular electrophysiology metadata table hierarchy
 
     Illustration of the hierarchy of metadata tables used to describe the organization of
     intracellular electrophysiology experiments.
+
+Recordings of intracellular electrophysiology stimuli and responses are represented
+with subclasses of :py:class:`~pynwb.icephys.PatchClampSeries` using the
+:py:class:`~pynwb.icephys.IntracellularElectrode` and :py:class:`~pynwb.device.Device`
+type to describe the electrode and device used.
+
+To describe the organization of intracellular experiments, the metadata is organized
+hierarchically in a sequence of tables. All of the tables are so-called DynamicTables
+enabling users to add columns for custom metadata.
 
 - :py:class:`~pynwb.icephys.IntracellularRecordingsTable` relates electrode, stimulus
   and response pairs and describes metadata specific to individual recordings.
@@ -52,6 +56,22 @@ enabling users to add columns for custom metadata.
 - :py:class:`~pynwb.icephys.ExperimentalConditionsTable` groups repetitions of
   intracellular recording from the :py:class:`~pynwb.icephys.RepetitionsTable`
   together that belong to the same experimental conditions.
+
+Storing data in hierarchical tables has the advantage that it allows us to avoid duplication of
+metadata. E.g., for a single experiment we only need to describe the metadata that is constant
+across an experimental condition as a single row in the :py:class:`~pynwb.icephys.ExperimentalConditionsTable`
+without having to replicate the same information across all repetitions and sequential-, simultaneous-, and
+individual intracellular recordings. For analysis, this means that we can easily focus on individual
+aspects of an experiment while still being able to easily access information about information from
+related tables.
+
+.. note:: All of the above mentioned metadata tables are optional and are created automatically
+          by the :py:class:`~pynwb.file.NWBFile` class the first time data is being
+          added to a table via the corresponding add functions. However, as tables at higher
+          levels of the hierarchy link to the other tables that are lower in the hierarchy,
+          we may only exclude tables from the top of the hierarchy. This means, for example,
+          a file containing a :py:class:`~pynwb.icephys.SimultaneousRecordingsTable` then
+          must also always contain a corresponding :py:class:`~pynwb.icephys.IntracellularRecordingsTable`.
 '''
 
 #######################
@@ -64,7 +84,9 @@ from datetime import datetime
 from dateutil.tz import tzlocal
 import numpy as np
 import pandas
-pandas.set_option("display.max_colwidth", 30)  # Set Pandas rendering option to avoid very wide cells
+# Set pandas rendering option to avoid very wide tables in the html docs
+pandas.set_option("display.max_colwidth", 30)
+pandas.set_option("display.max_rows", 10)
 
 # Import main NWB file class
 from pynwb import NWBFile
@@ -76,14 +98,17 @@ from pynwb import NWBHDF5IO
 from pynwb.core import DynamicTable, VectorData
 
 #######################
-# A brief example script
-# ------------------------------
+# A brief example
+# ---------------
 #
-# The following shows a brief example script, providing a brief overview of the
-# main steps to create and NWB icephys file. A more detailed dicussion of the
-# various steps is then provided afterwards. To avoid collisions between this
-# example script and the more detailed discussion we prefix
-# all variables in the example script with ``ex_``.
+# The following brief example provides a quick overview of the main steps
+# to create an NWBFile for intracelluar electrophysiology data. We then
+# discuss the individual steps in more detail afterwards.
+#
+# .. note:
+#    To avoid collisions between this example script and the more
+#    detailed discussion we prefix all variables in the example
+#    script with ``ex_``.
 
 # Create an ICEphysFile
 ex_nwbfile = NWBFile(session_description='my first recording',
@@ -128,7 +153,7 @@ ex_ir_index = ex_nwbfile.add_intracellular_recording(electrode=ex_electrode,
                                                      stimulus=ex_stimulus,
                                                      response=ex_response)
 
-# (B) Add a list of sweeps to the sweeps table
+# (B) Add a list of sweeps to the simultaneous recordings table
 ex_sweep_index = ex_nwbfile.add_icephys_simultaneous_recording(recordings=[ex_ir_index, ])
 
 # (C) Add a list of simultaneous recordings table indices as a sequential recording
@@ -143,7 +168,7 @@ run_index = ex_nwbfile.add_icephys_repetition(sequential_recordings=[ex_sequence
 ex_nwbfile.add_icephys_experimental_condition(repetitions=[run_index, ])
 
 # Write our test file
-ex_testpath = "ex_test_icephys_file.h5"
+ex_testpath = "ex_test_icephys_file.nwb"
 with NWBHDF5IO(ex_testpath, 'w') as io:
     io.write(ex_nwbfile)
 
@@ -168,8 +193,8 @@ except ImportError:  # ignore in case hdmf_docutils is not installed
 
 
 #######################
-# Creating an icephys NWB file
-# ------------------------------
+# Creating an NWB file for Intracellular electrophysiology
+# --------------------------------------------------------
 #
 # When creating an NWB file, the first step is to create the :py:class:`~pynwb.file.NWBFile`. The first
 # argument is is a brief description of the dataset.
@@ -278,10 +303,6 @@ rowindex = nwbfile.add_intracellular_recording(electrode=electrode,
                                                id=10)
 
 #######################
-# .. note:: All of the above mentioned metadata tables are optional and are created automatically
-#          by the :py:class:`~pynwb.file.NWBFile` class the first time data is being
-#          added to a table via the corresponding add functions.
-#
 # .. note:: Since :py:meth:`~pynwb.file.NWBFile.add_intracellular_recording` can automatically add
 #          the objects to the NWBFile  we do not need to separately call
 #          :py:meth:`~pynwb.file.NWBFile.add_stimulus` and :py:meth:`~pynwb.file.NWBFile.add_acquistion`
@@ -315,7 +336,10 @@ rowindex2 = nwbfile.add_intracellular_recording(electrode=electrode,
 # .. note:: A recording may optionally also consist of just an electrode and stimulus or electrode
 #           and response, but at least one of stimulus or response are required. If either stimulus or response
 #           are missing, then the stimulus and response are internally set ot the same TimeSeries and the
-#           start_index and index_count for the missing parameter are set to -1
+#           start_index and index_count for the missing parameter are set to -1. When retrieving
+#           data from the :py:class:`~pynwb.base.TimeSeriesReferenceVectorData` the missing values
+#           will be represented via masked numpy arrays, i.e., as masked values in a
+#           ``numpy.ma.masked_array`` or as a ``np.ma.core.MaskedConstant``.
 
 rowindex3 = nwbfile.add_intracellular_recording(electrode=electrode,
                                                 response=response,
@@ -537,7 +561,7 @@ rowindex = nwbfile.add_icephys_experimental_condition(
 #
 
 # Write our test file
-testpath = "test_icephys_file.h5"
+testpath = "test_icephys_file.nwb"
 with NWBHDF5IO(testpath, 'w') as io:
     io.write(nwbfile)
 
@@ -551,14 +575,17 @@ with NWBHDF5IO(testpath, 'r') as io:
 # -----------------------------
 #
 # All of the icephys metadata tables are available as attributes on the NWBFile object.
-# Below we simply convert the tables to Pandas dataframes to show their content.
-
+# For display purposed we convert the tables to pandas DataFrames to show their content.
+# For a more in-depth discussion of how to access and use the tables
+# see the tutorial on :ref:`icephys_pandas_tutorial`.
+pandas.set_option("display.max_columns", 6)  # avoid oversize table in the html docs
 nwbfile.intracellular_recordings.to_dataframe()
 
 #######################
 #
 
 # optionally we can ignore the id columns of the category subtables
+pandas.set_option("display.max_columns", 5)  # avoid oversize table in the html docs
 nwbfile.intracellular_recordings.to_dataframe(ignore_category_ids=True)
 
 #######################
@@ -580,24 +607,12 @@ nwbfile.icephys_repetitions.to_dataframe()
 
 nwbfile.icephys_experimental_conditions.to_dataframe()
 
-#######################
-# .. note:: As the tables in the hierarchy are optional, a given NWBFile may
-#       contain, e.g., a ``experimental_conditions`` table. To provide a consistent
-#       interface for users, and to avoid having to manually check which tables exist
-#       in a given file in order to find the "top-most" table in the hierarchy,
-#       NWBFile provides the ``:py:meth:`~pynwb.file.NWBFile.get_icephys_meta_parent_table``
-#       function. With this we can retrieve the highest table in the icephys metadata
-#       hierarchy that actually exists in the file. For example:
-
-top_table = nwbfile.get_icephys_meta_parent_table()
-print(top_table.neurodata_type)
-
 
 #######################
 # Validate data
 # ^^^^^^^^^^^^^
 #
-# This section is primarily for internal testing purpose to validate that the roundtrip
+# This section is for internal testing purposes only to validate that the roundtrip
 # of the data (i.e., generate --> write --> read) produces the correct results.
 
 # Read the data back in